% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_CI.R
\name{bootstrapped_CI}
\alias{bootstrapped_CI}
\title{Perform bootstrapping on variable}
\usage{
bootstrapped_CI(data_vector, n_resamples = 1000)
}
\arguments{
\item{data_vector}{observed data to sample from}

\item{n_resamples}{number of resamples to take}
}
\value{
A tibble containing mean of input and 95\% CI
}
\description{
Takes a numeric input, performs bootstrapping with a given number of resamples, 
computes 95\% Confidence Interval and returns a tibble with input mean and 
bootstrapped CI.
}
