% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{compare_groups}
\alias{compare_groups}
\title{Compare two subsets of factorial design cells}
\usage{
compare_groups(model, higher, lower, alpha = 0.05)
}
\arguments{
\item{model}{An object of class \code{\link[brms:brmsfit-class]{brmsfit}}.}

\item{higher}{'Higher' group specification.}

\item{lower}{'Lower' group specification.}

\item{alpha}{A single value (0, 1) as the alpha-level of the group 
comparison (defaults to 0.05).}
}
\value{
A list with summary statistics of the group comparison.
}
\description{
Convenience function for addressing directional hypotheses in factorial 
designs. It takes a \code{\link[brms:brm]{brms}} model fit, two group
specifications (two subsets of design cells), and the alpha-level of the 
comparison. It outputs the posterior mean of the 'higher' minus the 'lower' 
group, its 1-\code{alpha} percent credible interval, as well as the posterior 
probability and odds that the mean estimate of the 'higher' group is higher 
than that of the 'lower' group. A comparison of one group against the grand mean 
can be obtained by leaving out one of the two group specifications in the function 
call.
}
\examples{
library(brms)
# linear mixed effects model 
# regressing voice pitch against gender and context
m <- brm(formula = pitch ~ gender * context + (1 | subject + sentence), 
        data = data_polite)

# compare female speakers in informal contexts against male speakers in polite contexts
compare_groups(
 model  = m, 
 higher = gender == "F" & context == "inf",
 lower  = gender == "M" & context == "pol"
)

# compare informal contexts against polite contexts, averaged over gender
compare_groups(
 model  = m, 
 higher = context == "inf",
 lower  = context == "pol"
)

# compare female speakers against the grand mean
compare_groups(
 model  = m, 
 higher = gender == "F"
)

}
