% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cripa_nimble.R
\name{cripa_nimble}
\alias{cripa_nimble}
\title{Nimble CRIPA}
\usage{
cripa_nimble(
  seed,
  model_code,
  constants,
  data,
  inits,
  monitors,
  nburnin = 0,
  thin = 1,
  niter = 1e+05,
  nchains = 4
)
}
\arguments{
\item{seed}{set seed (maybe; I have to look into it)}

\item{model_code}{either a string containing the code of the model or the variable in which the code is stored}

\item{constants}{ALL the constants}

\item{data}{the dataset to be passed to `nimble::nimbleModel()`}

\item{inits}{inits parameter to be passed to `nimble::nimbleModel()`}

\item{monitors}{user modified monitors; leave empty for default (?)}

\item{nburnin}{...}

\item{thin}{...}

\item{niter}{number of iterations}

\item{nchains}{number of chains}
}
\value{
list
}
\description{
Note by MF: "Complile and Run individually for parallel assembly"
}
