% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisp_nimble.R
\name{crisp_nimble}
\alias{crisp_nimble}
\title{Nimble CRISP}
\usage{
crisp_nimble(
  model_code,
  constants,
  data,
  inits,
  monitors = c(),
  nburnin = 0,
  thin = 1,
  niter = 1e+05,
  nchains = 4,
  nclusters = 4
)
}
\arguments{
\item{model_code}{either a string containing the code of the model or the variable in which the code is stored}

\item{constants}{ALL the constants}

\item{data}{the dataset to be passed to `nimble::nimbleModel()`}

\item{inits}{inits parameter to be passed to `nimble::nimbleModel()`}

\item{monitors}{user modified monitors; leave empty for default (?)}

\item{nburnin}{...}

\item{thin}{...}

\item{niter}{number of iterations}

\item{nchains}{number of chains}

\item{nclusters}{number of clusters}
}
\value{
mcmc.list
}
\description{
Note by MF: 
Complile and Run In Sequence or Parallel
following template from https://r-nimble.org/nimbleExamples/parallelizing_NIMBLE.html
 "\[E]nsure that all NIMBLE execution, including model building, is conducted inside the 
  parallelized code. This ensures that all models and algorithms are independent objects
  that don't interfere with each other."
}
