% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_nimble.R
\name{crs_nimble}
\alias{crs_nimble}
\title{Nimble CRS}
\usage{
crs_nimble(
  model_code,
  constants,
  data,
  inits,
  monitors = c(),
  nburnin = 0,
  thin = 1,
  niter = 1e+05,
  nchains = 4
)
}
\arguments{
\item{model_code}{either a string containing the code of the model or the variable in which the code is stored}

\item{constants}{ALL the constants}

\item{data}{the dataset to be passed to `nimble::nimbleModel()`}

\item{inits}{inits parameter to be passed to `nimble::nimbleModel()`}

\item{monitors}{user modified monitors; leave empty if}

\item{nburnin}{...}

\item{thin}{...}

\item{niter}{number of iterations}

\item{nchains}{number of chains}
}
\value{
mcmc.list
}
\description{
convenience function for building and compiling numble models (I guess? Maybe?)
Note by MF: "Compile and Run"
}
