% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_avocado_raw}
\alias{data_avocado_raw}
\title{Avocado prices}
\format{
A data frame with 18249 rows and 14 variables.
The most important variables in this data set are:
\describe{
  \item{Date}{Date of the observation.}
  \item{AveragePrice}{Average price of a single avocado.}
  \item{Total Volume}{Total number of avocados sold.}
  \item{type}{Whether the price/amount is for conventional 
  or organic avocados.}
  \item{4046}{Total number of small avocados sold (PLU 4046).}
  \item{4225}{Total number of medium avocados sold (PLU 4225).}
  \item{4770}{Total number of large avocados sold (PLU 4770).}
}
}
\source{
\url{https://www.kaggle.com/neuromusic/avocado-prices}
}
\usage{
data_avocado_raw
}
\description{
This data set was downloaded from 
\href{https://www.kaggle.com/}{Kaggle}. It includes 
information about the prices of (Hass) avocados and 
the amount sold (of different kinds) at different 
points in time.
}
\seealso{
\code{data_avocado} for a preprocessed version of this data set.
 
\href{https://michael-franke.github.io/intro-data-analysis/app-93-data-sets-avocado.html}{The web-book chapter that covers the Avocado data set}.
}
\keyword{datasets}
