% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_BLJM_raw}
\alias{data_BLJM_raw}
\title{Bio-Logic Jazz-Metal}
\format{
A data frame with 306 rows and 19 variables.
The most important variables in this data set are:
\describe{
  \item{submission_id}{A unique identifier for each participant.}
  \item{option 1}{What the choice options were.}
  \item{option 2}{What the choice options were.}
  \item{response}{Which of the two options was chosen.}
}
}
\source{
\url{https://raw.githubusercontent.com/michael-franke/intro-data-analysis/master/data_sets/bio-logic-jazz-metal-data-raw.csv}
}
\usage{
data_BLJM_raw
}
\description{
This data set is from a very short and 
non-serious online experiment that asked for just three 
binary decisions from each participant, namely their 
spontaneous preference for one of two presented options 
(biology vs. logic, jazz vs. metal, and mountains vs. beach).
}
\seealso{
\code{data_BLJM} for a preprocessed version of this 
data set.

\href{https://michael-franke.github.io/intro-data-analysis/app-93-data-sets-BLJM.html}{The web-book chapter that covers the BLJM data set}.
}
\keyword{datasets}
