% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_MC_raw}
\alias{data_MC_raw}
\title{Mental Chronometry}
\format{
A data frame with 3750 rows and 32 variables. 
The most important variables in this data set are:
\describe{
  \item{submission_id}{A unique identifier for each participant.}
  \item{RT}{The reaction time for each trial.}
  \item{response}{Response to the presented stimulus in the 
  reaction or go/no-go condition.}
  \item{key_pressed}{The key that was pressed in the discrimination 
  condition (f or j).}
  \item{trial_type}{The condition (reaction, go/no-go, discrimination), 
  as well as whether it was a practice or main trial.}
  \item{stimulus}{The stimulus displayed on the screen; either a 
  blue circle or a blue square.}
}
}
\source{
\url{https://raw.githubusercontent.com/michael-franke/intro-data-analysis/master/data_sets/mental-chrono-data_raw.csv}
}
\usage{
data_MC_raw
}
\description{
This data set is from an online experiment 
investigating differences in reaction times and accuracy 
in solving three tasks of increasing complexity.
}
\seealso{
\code{data_MC_preprocessed} for a preprocessed version of this data set.

\code{data_MC_cleaned} for a cleaned version of this data set.

\href{https://michael-franke.github.io/intro-data-analysis/app-93-data-sets-mental-chronometry.html}{The web-book chapter that covers the MC data set}.
}
\keyword{datasets}
