% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_murder}
\alias{data_murder}
\title{Murder data - fictitious}
\format{
A data frame with 20 rows and 4 variables. Each row 
in this data set shows data from a city. The variables are:
\describe{
  \item{murder_rate}{Annual murder rate per million inhabitants.}
  \item{low_income}{Percentage of inhabitants with a low income 
  (however that is defined).}
  \item{unemployment}{Percentage of unemployed inhabitants.}
  \item{population}{Number of inhabitants of a city.}
}
}
\source{
\url{https://raw.githubusercontent.com/michael-franke/intro-data-analysis/master/data_sets/murder_rates.csv}
}
\usage{
data_murder
}
\description{
The murder data set contains information 
about the relative number of murders in American cities. 
It also contains further socio-economic information, such 
as a city’s unemployment rate, and the percentage of 
inhabitants with a low income. This data set should only 
be used for illustration. No further real-world conclusions 
should be drawn from the data, as it is entirely fictitious.
}
\seealso{
\href{https://michael-franke.github.io/intro-data-analysis/app-93-data-sets-murder-data.html}{The web-book chapter that covers the Murder data set}.
}
\keyword{datasets}
