% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_polite}
\alias{data_polite}
\title{Politeness data}
\format{
A data frame with 83 rows and 5 variables.
The most important variables in this data set are:
\describe{
  \item{subject}{A unique identifier for each participant.}
  \item{gender}{Gender of the participant; male or female.}
  \item{context}{Linguistic context; informal or polite.}
  \item{pitch}{Voice pitch of the speaker.}
}
}
\source{
\url{https://raw.githubusercontent.com/michael-franke/bayes_mixed_regression_tutorial/master/code/politeness_data.csv}
}
\usage{
data_polite
}
\description{
This data set stems from a tutorial on 
Bayesian regression modeling for factorial designs by 
Franke and Roettger (2019), who preprocessed the data 
from Winter and Grawunder (2012). The data set includes 
voice pitch of male and female speakers in both polite 
and informal linguistic contexts.
}
\references{
Franke, M., Roettger, T. B. (2019). 
Bayesian regression modeling (for factorial designs): A tutorial.

Winter, B., Grawunder, S. (2012). 
The phonetic profile of Korean formality. 
Journal of Phonetics, 40, 808-815.
}
\keyword{datasets}
