% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ST_raw}
\alias{data_ST_raw}
\title{Simon Task}
\format{
A data frame with 25560 rows and 15 variables.
The most important variables in this data set are:
\describe{
  \item{submission_id}{A unique identifier for each participant.}
  \item{RT}{The reaction time for each trial.}
  \item{condition}{Whether the trial was a congruent or 
  an incongruent trial.}
  \item{correctness}{Whether the answer in the current 
  trial was correct or incorrect.}
  \item{trial_type}{Whether the data is from a practice 
  or a main test trial.}
}
}
\source{
\url{https://raw.githubusercontent.com/michael-franke/intro-data-analysis/master/data_sets/simon-task.csv}
}
\usage{
data_ST_raw
}
\description{
This data set is from an online experiment 
in which participants classified a shape presented on the 
screen as a square or circle. The shape was displayed on 
either the left or right side of the screen.
}
\seealso{
\code{data_ST} for a cleaned version of this data set. 

\href{https://michael-franke.github.io/intro-data-analysis/app-93-data-sets-simon-task.html}{The web-book chapter that covers the Simon Task data set}.
}
\keyword{datasets}
