% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_WorldTemp}
\alias{data_WorldTemp}
\title{Annual average world temperature}
\format{
A data frame with 269 rows and 4 variables.
\describe{
  \item{year}{Year of the observation (1750-2019).}
  \item{anomaly}{Deviation from the grand mean of 
  1750-1980, which equals 8.61 degrees Celcius.}
  \item{uncertainty}{Measure of uncertainty associated 
  with the reported anomaly.}
  \item{avg_temp}{The annual average world surface 
  temperature.}
}
}
\source{
\url{https://raw.githubusercontent.com/michael-franke/intro-data-analysis/master/data_sets/average-world-temperature.csv}
}
\usage{
data_WorldTemp
}
\description{
This data set was downloaded from 
\href{http://berkeleyearth.org/}{Berkeley Earth} on 
October 6th, 2020. Specifically, it contains the time 
series data for “land only” using the annual summary 
of monthly average temperature. We have added to the 
data set used here the absolute average temperature. 
(Berkeley Earth only lists the “annual anomaly”, i.e., 
the deviation from a grand mean.)
}
\seealso{
\href{http://berkeleyearth.org/data-new/}{The origin and composition of this data set}.

\href{https://michael-franke.github.io/intro-data-analysis/app-93-data-sets-temperature.html}{The web-book chapter that covers the World Temperature data set}.
}
\keyword{datasets}
