% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{extract_draws}
\alias{extract_draws}
\title{Extract posterior draws for one subset of factorial design cells}
\usage{
extract_draws(model, ...)
}
\arguments{
\item{model}{An object of class \code{\link[brms:brmsfit-class]{brmsfit}}.}

\item{...}{The group specification.}
}
\value{
A data frame containing posterior draws for the specified group.
}
\description{
This function takes as input a \code{\link[brms:brm]{brms}} model fit for a 
factorial design and a group specification (one subset of the design cells). It
returns the posterior draws for that group. If no group is specified, the returned 
posterior draws are grand means.
}
\examples{
library(brms)
# linear mixed effects model 
# regressing voice pitch against gender and context
m <- brm(formula = pitch ~ gender * context + (1 | subject + sentence), 
        data = data_polite)

# extract posterior draws of female speakers in informal contexts
extract_draws(m, gender == "F", context == "inf")
extract_draws(m, gender == "F" & context == "inf")

# extract posterior draws of male speakers or informal contexts
extract_draws(m, gender == "M" | context == "inf")

# averaged over gender, extract posterior draws for all but polite contexts
extract_draws(m, context != "pol")

# extract the posterior draws averaged over all factors
extract_draws(m)
}
