% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{get_cell_definitions}
\alias{get_cell_definitions}
\title{Obtain information about factors in regression models}
\usage{
get_cell_definitions(model)
}
\arguments{
\item{model}{An object of class \code{\link[brms:brmsfit-class]{brmsfit}}.}
}
\value{
A \code{\link[tibble:tibble-package]{tibble}} containing the minimal
design matrix.
}
\description{
This function takes a \code{\link[brms:brm]{brms}} model fit for a 
factorial design and returns all design cells along with their encoding in 
the regression model.
}
\examples{
library(brms)
# linear mixed effects model 
# regressing voice pitch against gender and context
m <- brm(formula = pitch ~ gender * context + (1 | subject + sentence), 
        data = data_polite)
get_cell_definitions(m)
}
