% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples_regression_noninformative.R
\name{get_samples_regression_noninformative}
\alias{get_samples_regression_noninformative}
\title{Get posterior samples from a non-informative priors regression model}
\usage{
get_samples_regression_noninformative(X, y, n_samples = 1000)
}
\arguments{
\item{X}{design matrix}

\item{y}{dependent variable}

\item{n_samples}{number of samples taken from the posterior}
}
\value{
A tibble of sample triplets.
}
\description{
Convenience function for obtaining posterior samples from a 
non-informative priors linear regression model for inferring 
the parameters \code{intercept}, \code{slope} and \code{sigma}.
}
