% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples_single_noninformative.R
\name{get_samples_single_noninformative}
\alias{get_samples_single_noninformative}
\title{Get posterior samples from a non-informative priors model}
\usage{
get_samples_single_noninformative(data_vector, n_samples = 1000)
}
\arguments{
\item{data_vector}{vector of metric observations}

\item{n_samples}{number of samples taken from the posterior}
}
\value{
A tibble of associated samples.
}
\description{
Convenience function for obtaining posterior samples from 
a non-informative priors model for inferring the parameters 
\code{mu} and \code{sigma} of a normal distribution.
}
