% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples_single_normal_conjugate.R
\name{get_samples_single_normal_conjugate}
\alias{get_samples_single_normal_conjugate}
\title{Get posterior samples from a conjugate priors model}
\usage{
get_samples_single_normal_conjugate(
  data_vector,
  nu = 1,
  var = 1,
  mu = 0,
  kappa = 1,
  n_samples = 1000
)
}
\arguments{
\item{data_vector}{vector of metric observations}

\item{nu}{prior on nu}

\item{var}{prior on the variance}

\item{mu}{prior on the mean}

\item{kappa}{prior on kappa}

\item{n_samples}{number of samples taken from the posterior}
}
\value{
A tibble of associated samples.
}
\description{
Convenience function for obtaining posterior samples 
from a conjugate priors model for inferring the parameters 
\code{mu} and \code{sigma} of a normal distribution.
}
