% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_Norm_inv_chisq.R
\name{sample_Norm_inv_chisq}
\alias{sample_Norm_inv_chisq}
\title{Get samples from the ‘normal inverse-chi^2' prior}
\usage{
sample_Norm_inv_chisq(n_samples = 10000, nu = 1, var = 1, mu = 0, kappa = 1)
}
\arguments{
\item{n_samples}{number of samples taken from the prior}

\item{nu}{prior on nu}

\item{var}{prior on the variance}

\item{mu}{prior on the mean}

\item{kappa}{prior on kappa}
}
\value{
A tibble of associated samples.
}
\description{
Convenience function for sampling from the prior 
distribution over \code{mu} and \code{sigma}.
}
