% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_mcmc_list.R
\name{summarize_mcmc_list}
\alias{summarize_mcmc_list}
\title{Summarize mcmc.list objects}
\usage{
summarize_mcmc_list(samples)
}
\arguments{
\item{samples}{mcmc.list to summarize}
}
\value{
tibble
}
\description{
Summarize mcmc.list objects
}
