% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aida.R
\name{theme_aida}
\alias{theme_aida}
\title{ggplot2 theme for the AIDA course}
\usage{
theme_aida(
  title.size = 16,
  text.size = 14,
  legend.position = "top",
  show.axis = FALSE,
  plot.margin = c(0.2, 0.1, 0.2, 0.1)
)
}
\arguments{
\item{title.size}{title element size in pts}

\item{text.size}{text element size in pts}

\item{legend.position}{either one of "none", "left", "right", "bottom", "top"}

\item{show.axis}{boolean}

\item{plot.margin}{margin around entire plot (unit with the sizes of the top, right, bottom, and left margins)}
}
\description{
ggplot2 theme for the AIDA course
}
