% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_polite}
\alias{data_polite}
\title{Politeness data}
\format{
A data frame with 83 rows and 5 variables:
\describe{
  \item{subject}{A unique identifier for each participant.}
  \item{gender}{An indicator of each participants gender 
  (only binary).}
  \item{sentence}{An indicator of the sentence spoken by 
  the participant.}
  \item{context}{The main manipulation of whether the context 
  was a “polite” or “informal” setting.}
  \item{pitch}{The measured voice pitch (presumably: average 
  over the sentence spoken).}
}
}
\source{
\url{https://raw.githubusercontent.com/michael-franke/intro-data-analysis/master/data_sets/politeness_data.csv}
}
\usage{
data_polite
}
\description{
This data set is a preprocessed and shortened
version of the data provided by Winter and Grawunder (2012). 
The data set includes voice pitch of male and female Korean 
speakers in both polite and informal linguistic contexts.
}
\references{
Winter, B., Grawunder, S. (2012). 
The phonetic profile of Korean formality. 
Journal of Phonetics, 40, 808-815.
}
\keyword{datasets}
