% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sample_vector.R
\name{summarize_sample_vector}
\alias{summarize_sample_vector}
\title{Summarize single vector of samples (lower/upper 95% Credible Interval & mean)}
\usage{
summarize_sample_vector(samples, name = "")
}
\arguments{
\item{samples}{numeric vector of samples}
}
\value{
tibble
}
\description{
Summarize single vector of samples (lower/upper 95% Credible Interval & mean)
}
